separate provider from client state
We want to test using different devices on the same mailbox. We are also
looking into using a transport mechanism that works between different
devices and browser tabs.

In preperation for this separate the messages stored on the server from
the client state. On switching the user we now go through the inbox and
recover the autocrypt state just like we would when scanning an inbox on
a new device.

There are two implementations for the provider side right now. One is
based on a simple object, the other one uses localstorage. The former
behaves as before. The latter will keep mails on the 'server side' on a
page refresh.

In general i am trying to map the actual entities in a mail system to
data structures and also have information flow in a similar manner.